requirements.txt
django==1.11
django-rest-framework
django-filter>=1.0.1
